requirements.txt
streamlit
faiss-cpu
sentence-transformers
pypdf
numpy
requests
tqdm
